+++ pentobi/qml/ToolBar.qml
@@ -282,9 +282,6 @@ Item {
                     id: menuComponent
 
                     Pentobi.Menu {
-                        dynamicWidth: false
-                        width: Math.min(font.pixelSize * (isDesktop ? 11 : 18),
-                                        rootWindow.contentItem.width)
                         closePolicy: Popup.CloseOnPressOutsideParent
                                      | Popup.CloseOnEscape